Change of structures, design and setup of other pages within Labs
Made extensive changes to the inheritance of liquid templates. Changed
base fonts to fit current design. Added rough text for other pages.
Changed the lookup of dataset list items to come from _data folder —
this is probably a good model for other lists, reserving the posts
functionality for a blog.